Infra Updates - Working on Lambda Deployments
- Renamed scheular to scheduler
- Reviewed/updated build/package/config/deploy
