+++ cla-backend-go/v2/signatures/converters.go
@@ -1,0 +1,45 @@
+// Copyright The Linux Foundation and each contributor to CommunityBridge.
+// SPDX-License-Identifier: MIT
+
+package signatures
+
+import (
+	v1Models "github.com/communitybridge/easycla/cla-backend-go/gen/models"
+	"github.com/communitybridge/easycla/cla-backend-go/gen/v2/models"
+	"github.com/jinzhu/copier"
+)
+
+func v2Signature(src *v1Models.Signature) (*models.Signature, error) {
+	var dst models.Signature
+	err := copier.Copy(&dst, src)
+	if err != nil {
+		return nil, err
+	}
+	return &dst, nil
+}
+
+func v2Signatures(src *v1Models.Signatures) (*models.Signatures, error) {
+	var dst models.Signatures
+	err := copier.Copy(&dst, src)
+	if err != nil {
+		return nil, err
+	}
+	return &dst, nil
+}
+
+func v2SignaturesReplaceCompanyID(src *v1Models.Signatures, internalID, externalID string) (*models.Signatures, error) {
+	var dst models.Signatures
+	err := copier.Copy(&dst, src)
+	if err != nil {
+		return nil, err
+	}
+
+	// Replace the internal ID with the External ID
+	for _, sig := range dst.Signatures {
+		if sig.SignatureReferenceID == internalID {
+			sig.SignatureReferenceID = externalID
+		}
+	}
+
+	return &dst, nil
+}
