+++ cla-backend/serverless.yml
@@ -7,6 +7,9 @@ package:
     - ./backend-aws-lambda
     - ./metrics-aws-lambda
     - ./dynamo-events-lambda
+    - ./zipbuilder-scheduler-lambda
+    - ./zipbuilder-lambda
+    - ./functional-tests
     - dev.sh
     - docs/**
     - helpers/**
@@ -362,6 +365,34 @@ functions:
       individually: true
       include:
         - ./metrics-aws-lambda
+
+  zipbuilder-scheduler-lambda:
+    handler: zipbuilder-scheduler-lambda
+    name: ${self:service}-${opt:stage, self:provider.stage, 'dev'}-zipbuilder-scheduler-lambda
+    description: "call zipbuilder-lambda for all cla groups periodically"
+    runtime: go1.x
+    timeout: 900 # maximum time allowed
+    events:
+      - schedule:
+          description: 'build signatures pdf for cla group'
+          rate: rate(1 hour)
+          enabled: true
+    package:
+      individually: true
+      include:
+        - ./zipbuilder-scheduler-lambda
+
+  zipbuilder-lambda:
+    handler: zipbuilder-lambda
+    name: ${self:service}-${opt:stage, self:provider.stage, 'dev'}-zipbuilder-lambda
+    description: "build zip of signed signature pdf for cla group"
+    runtime: go1.x
+    timeout: 900 # maximum time allowed
+    memorySize: 1024
+    package:
+      individually: true
+      include:
+        - ./zipbuilder-lambda
 
   apiv1:
     handler: wsgi_handler.handler
